Fix exports and rename things
