+++ lib/components/Charts/exports.ts
@@ -1,1 +1,11 @@
-export * from "./AreaChart"
+import { Component } from "@/lib/component"
+
+import { AreaChart as AreaChartComponent } from "./AreaChart"
+
+export const AreaChart = Component(
+  {
+    name: "AreaChart",
+    type: "info",
+  },
+  AreaChartComponent
+)
