+++ lib/components/Insights/Charts/AreaChartInsight/index.stories.tsx
