+++ lib/components/Insights/Charts/AreaChartInsight/index.tsx
