+++ lib/components/Insights/Charts/storybook-utils.tsx
@@ -1,4 +1,4 @@
-import ContainerStory from "../Container/index.stories"
+import ContainerStory from "../InsightsContainer/index.stories"
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 const { children, ...containerStoryArgs } = ContainerStory.args
